+++ .eleventy.cjs
@@ -3,15 +3,16 @@ const anchorsPlugin = require('@orchidjs/eleventy-plugin-ids');
 const syntaxHighlight = require('@11ty/eleventy-plugin-syntaxhighlight');
 const directoryOutputPlugin = require('@11ty/eleventy-plugin-directory-output');
 
-const pfeAssetsPlugin = require('./docs/_plugins/pfe-assets.cjs');
-const customElementsManifestPlugin = require('./docs/_plugins/custom-elements-manifest.cjs');
-const orderTagsPlugin = require('./docs/_plugins/order-tags.cjs');
-const todosPlugin = require('./docs/_plugins/todos.cjs');
+const pfeAssetsPlugin = require('@patternfly/pfe-tools/11ty/plugins/pfe-assets.cjs');
+
+const customElementsManifestPlugin = require('@patternfly/pfe-tools/11ty/plugins/custom-elements-manifest.cjs');
+const orderTagsPlugin = require('@patternfly/pfe-tools/11ty/plugins/order-tags.cjs');
+const todosPlugin = require('@patternfly/pfe-tools/11ty/plugins/todos.cjs');
 
 const markdownIt = require('markdown-it');
 const markdownItAnchor = require('markdown-it-anchor');
 
-const pluginToc = require('eleventy-plugin-toc');
+const pluginToc = require('@patternfly/pfe-tools/11ty/plugins/table-of-contents.cjs');
 
 const markdownLib = markdownIt({ html: true })
   .use(markdownItAnchor);