+++ docs/_plugins/types.d.ts
@@ -1,1 +1,0 @@
-export type * as BuilderApi from 'mdast-builder';